third_party/blink/web_tests/external/wpt/custom-elements/registries/pseudo-class-defined.window-expected.txt

This is a testharness.js-based test.
[FAIL] "uncustomized" :defined doesn't care about your registry'
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] "custom" :defined doesn't care about your registry
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] pseudo-class-defined
  assert_true: expected true got false
Harness: the test ran to completion.
